# docs/swagger.yaml
basePath: /v2
definitions:
...
  commentdto.CommentAuthor:
    properties:
      author_id:
        type: integer
      bio:
        type: string
      profile_image:
        type: string
      username:
        type: string
    type: object
  commentdto.CreateCommentRequestDTO:
    properties:
      body:
        type: string
    type: object
  commentdto.CreateCommentResponseDTO:
    properties:
      author:
        $ref: '#/definitions/commentdto.CommentAuthor'
      body:
        type: string
      comment_id:
        type: integer
      created_at:
        type: string
    type: object
  commentdto.ReadCommentsResponseDTO:
    properties:
      author:
        $ref: '#/definitions/commentdto.CommentAuthor'
      body:
        type: string
      comment_id:
        type: integer
      created_at:
        type: string
      updated_at:
        type: string
    type: object
  commentdto.ReadCommentsResponseWrapperDTO:
    properties:
      comments:
        items:
          $ref: '#/definitions/commentdto.ReadCommentsResponseDTO'
        type: array
    type: object
  commentdto.UpdateCommentRequestDTO:
    properties:
      body:
        type: string
    type: object
  commentdto.UpdateCommentResponseDTO:
    properties:
      author:
        $ref: '#/definitions/commentdto.CommentAuthor'
      body:
        type: string
      comment_id:
        type: integer
      created_at:
        type: string
      updated_at:
        type: string
    type: object
  types.ErrorDetail:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  types.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/types.ErrorDetail'
    type: object
  userdto.ReadUserProfileResponseDTO:
    properties:
      bio:
        type: string
      profile_image:
        type: string
      username:
        type: string
    type: object
  userdto.ReadUserProfileResponseWrapperDTO:
    properties:
      user:
        $ref: '#/definitions/userdto.ReadUserProfileResponseDTO'
    type: object
  userdto.UpdateUserProfileRequestDTO:
    properties:
      bio:
        type: string
      email:
        type: string
      password:
        type: string
      profile_image:
        type: string
      username:
        type: string
    type: object
  userdto.UpdateUserProfileResponseDTO:
    properties:
      bio:
        type: string
      profile_image:
        type: string
      username:
        type: string
    type: object
  userdto.UpdateUserProfileResponseWrapperDTO:
    properties:
      user:
        $ref: '#/definitions/userdto.UpdateUserProfileResponseDTO'
    type: object
host: petstore.swagger.io
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /user/{user_id}/article/{article_id}/comment:
    post:
      consumes:
      - application/json
      description: 댓글을 생성합니다.
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: article author id(기사 소유자)
        in: path
        name: user_id
        required: true
        type: integer
      - description: article id(기사 ID)
        in: path
        name: article_id
        required: true
        type: integer
      - description: 댓글 내용
        in: body
        name: createCommentReq
        required: true
        schema:
          $ref: '#/definitions/commentdto.CreateCommentRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/commentdto.CreateCommentResponseDTO'
        "400":
          description: 잘못된 값을 요청한 경우
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: 요청을 처리하지 못한 경우
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: 네트워크 에러
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: 댓글 생성
      tags:
      - Comment tag
  /user/{user_id}/article/{article_id}/comment/{comment_id}:
    delete:
      consumes:
      - application/json
      description: 댓글을 삭제합니다.
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: article author id(기사 소유자)
        in: path
        name: user_id
        required: true
        type: integer
      - description: article id(기사 ID)
        in: path
        name: article_id
        required: true
        type: integer
      - description: comment id(댓글 ID)
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 정상적으로 삭제됨
          schema:
            type: string
        "400":
          description: 잘못된 값을 요청한 경우
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: 네트워크 에러
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: 댓글 삭제
      tags:
      - Comment tag
    put:
      consumes:
      - application/json
      description: 댓글을 수정합니다.
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: article author id(기사 소유자)
        in: path
        name: user_id
        required: true
        type: integer
      - description: article id(기사 ID)
        in: path
        name: article_id
        required: true
        type: integer
      - description: comment id(댓글 ID)
        in: path
        name: comment_id
        required: true
        type: integer
      - description: updateCommentReq
        in: body
        name: updateCommentReq
        required: true
        schema:
          $ref: '#/definitions/commentdto.UpdateCommentRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commentdto.UpdateCommentResponseDTO'
        "400":
          description: 잘못된 값을 요청한 경우
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: 댓글을 찾을 수 없는 경우
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: 요청을 처리하지 못한 경우
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: 네트워크 에러
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: 댓글 수정
      tags:
      - Comment tag
  /user/{user_id}/article/{article_id}/comments:
    get:
      consumes:
      - application/json
      description: 기사의 댓글 목록을 조회합니다.
      parameters:
      - description: article author id(기사 소유자)
        in: path
        name: user_id
        required: true
        type: integer
      - description: article id(기사 ID)
        in: path
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commentdto.ReadCommentsResponseWrapperDTO'
        "400":
          description: 잘못된 값을 요청한 경우
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: 네트워크 에러
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: 댓글 목록 조회
      tags:
      - Comment tag
  /user/{user_id}/profile:
    get:
      consumes:
      - application/json
      description: Read user profile
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/userdto.ReadUserProfileResponseWrapperDTO'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: network error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Read user profile
      tags:
      - User tag
    put:
      consumes:
      - application/json
      description: Update user profile
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: updateUserProfileReq
        in: body
        name: updateUserProfileReq
        required: true
        schema:
          $ref: '#/definitions/userdto.UpdateUserProfileRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/userdto.UpdateUserProfileResponseWrapperDTO'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: 요청을 제대로 수행하지 못함
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: network error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update user profile
      tags:
      - User tag
...
swagger: "2.0"